Finished the system for placing buildings, and also made it take into…

+++ Assets/Scripts/EconomyManager.cs
@@ -9,7 +9,7 @@ public class EconomyManager : MonoBehaviour
     //Effectable Values
     public int currentGold;
     public int currentSilver;
-    public int currentTechPoints;
+    public float currentTechPoints;
     public int currentWarWeariness;
     public int currentLoyalty; // Default Value is 100
